+++ frontends/web/src/routes/account/add/add-account.tsx
@@ -78,7 +78,6 @@ const AddAccountSteps = ({
         }} />
         <Input
           autoFocus
-          className={styles.accountNameInput}
           ref={inputRef}
           id="accountName"
           onInput={onAccountNameInput}