+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Portfolio: reduce waiting time to display the chart
 - iOS: Add haptic feedback on toggle component
 - Apply rounded corners in several places and components of the app
+- Send: enable rotating currencies in send window.
+- Show the fee's fiat value in send-to-self txs for consistency with others, and include the send-to-self amount in the left-side label
 
 ## v4.50.1
 - Fix a bug that would delay showing watch-only accounts.
@@ -24,6 +26,7 @@
 - fix language sometimes not persistent across app restarts
 - Android: fix layout issues with status and navigation bars.
 - Android: make the UI work with responsive font sizes and adhere to OS font size settings
+- Android: prompt to save exported files instead of sharing
 - Add "Change device password" functionality (in Settings)
 - Add icons for CTA and action buttons in account page
 - Restructure "Manage device" tab in settings